+++ plugins/publish/publish/src/main/kotlin/PublishAccessors.kt
@@ -1,0 +1,11 @@
+@file:Suppress("FunctionName", "TopLevelPropertyNaming", "ObjectPropertyName")
+
+import org.gradle.plugin.use.PluginDependenciesSpec
+import org.gradle.plugin.use.PluginDependencySpec
+
+val PluginDependenciesSpec.`javiersc-publish`: PluginDependencySpec
+    get() = javiersc("publish")
+
+fun PluginDependenciesSpec.`javiersc-publish`(
+    version: String,
+): PluginDependencySpec = javiersc("publish", version)