+++ plugins/publish/publish/src/main/kotlin/internal/JarTasks.kt
@@ -1,0 +1,54 @@
+@file:Suppress("PackageDirectoryMismatch")
+
+package com.javiersc.gradle.plugins.publish.internal
+
+import com.android.build.gradle.LibraryExtension
+import com.javiersc.gradle.plugins.publish.isAndroidLibrary
+import com.javiersc.gradle.plugins.publish.isGradlePlugin
+import com.javiersc.gradle.plugins.publish.isKotlinJvm
+import com.javiersc.gradle.plugins.publish.isKotlinMultiplatformWithAndroid
+import com.javiersc.gradle.plugins.publish.isVersionCatalog
+import java.io.File
+import org.gradle.api.Project
+import org.gradle.api.tasks.SourceSetContainer
+import org.gradle.api.tasks.bundling.Jar
+import org.gradle.kotlin.dsl.create
+import org.gradle.kotlin.dsl.get
+
+internal val Project.docsJar: Jar
+    get() =
+        tasks.create<Jar>("docsJar") {
+            group = "build"
+            description = "Assembles Javadoc jar file from for publishing"
+            archiveClassifier.set("javadoc")
+        }
+
+internal val Project.sourcesJar: Jar
+    get() =
+        tasks.create<Jar>("sourcesJar") {
+            group = "build"
+            description = "Assembles Sources jar file for publishing"
+            archiveClassifier.set("sources")
+
+            val sources: Iterable<File> =
+                when {
+                    isKotlinMultiplatformWithAndroid -> emptySet()
+                    isKotlinJvm || isGradlePlugin -> {
+                        (project.properties["sourceSets"] as SourceSetContainer)["main"].allSource
+                    }
+                    isAndroidLibrary -> {
+                        (project.extensions.getByName("android") as LibraryExtension)
+                            .sourceSets
+                            .named("main")
+                            .get()
+                            .java
+                            .srcDirs
+                    }
+                    isGradlePlugin -> {
+                        (project.properties["sourceSets"] as SourceSetContainer)["main"].allSource
+                    }
+                    isVersionCatalog -> emptySet()
+                    else -> emptySet()
+                }
+            if (sources.toList().isNotEmpty()) from(sources)
+        }